backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.1.3
